refactor(minesweeper) : PositionIdFactory 이름 변경
