+++ src/main/kotlin/domain/MineFactory.kt
@@ -2,13 +2,13 @@ package domain
 
 import model.GameData
 
-class MineFactory(private val randomPositionIdFactory: RandomPositionIdFactory = DefaultRandomPositionIdFactory()) {
+class MineFactory(private val positionIdFactory: PositionIdFactory = RandomPositionIdFactory()) {
 
     fun createMines(gameData: GameData): Mines {
 
         val maxPositionId = gameData.width * gameData.height - 1
 
-        val mines = randomPositionIdFactory.positionIds(gameData.mineNumber, maxPositionId).map {
+        val mines = positionIdFactory.positionIds(gameData.mineNumber, maxPositionId).map {
             Mine(it.position(gameData.width))
         }
