+++ src/main/kotlin/domain/PositionIdFactory.kt
@@ -1,0 +1,5 @@
+package domain
+
+interface PositionIdFactory {
+    fun positionIds(count: Int, positionIdMax: Int): List<PositionId>
+}
